+++ web/src/util/serialization/apd.js
@@ -36,7 +36,7 @@ export const fromAPI = apdAPI => {
       ({
         contractorResources,
         expenses,
-        goals,
+        objectives,
         schedule,
         statePersonnel,
         ...activity
@@ -57,10 +57,14 @@ export const fromAPI = apdAPI => {
           ...e
         })),
 
-        goals: goals.map(g => ({
-          ...g,
+        objectives: objectives.map(o => ({
+          ...o,
           initialCollapsed: true,
-          key: generateKey()
+          key: generateKey(),
+          keyResults: o.keyResults.map(kr => ({
+            key: generateKey(),
+            ...kr
+          }))
         })),
 
         schedule: schedule.map(s => ({